factored out generic Dynamics class from ODEDynamics and moved torchdiffeq specific implementations out of ODEDynamics class
